add ngen.config.path_pair subpackage
The goal of this subpackage is to provide classes that encapsulate a
`pathlib.Path` and so `T` or collection of `T`'s and flexibility for
writing, reading, serializing, deserializing the inner `T` / `T`'s.

This functionality is desirable in the context of pydantic, where you
want a field that serializes as a Path but holds onto an inner
deserialized T (this could be another pydantic model for example).