chore: use pyproject.toml; track version w/ setuptools_scm
remove setup.cfg in favor of pyproject.toml files.
this consolidates metadata and allows us to use...
setuptools_scm to version all packages.
previously a namespace package would contain an _version.py file,
w/ __version__ = "x.y.z".
this was a little cumbersome, but worse, there was not a good way to
determine which version of the code someone was using.

setuptools_scm will now automatically create ._version using git tags
when the package is installed. additionally, if the installed package is
not from a specific tag, setuptools_scm will add metadata to the version
including the code's git hash and date of install. this in-theory will
vastly improve our ability to reproduce user submitted issues.

tags follow the following semantics: ngen.<package-name>/v<x>.<y>.<z>
tags that do not follow this format will not be recognized and used by
setuptools_scm to create the correct pkg version.