feat: value_unit_pair module
Add `ValueUnitPair` and `ListUnitPair` classes. The former captures the
concept of a value and a unit (i.e. `42[meters]`) while the latter
is an extension of the concept but for a list of values.

Why do we need this? BMI modules: soil freeze thaw, soil moisture
profile, lgar, and CFE all use this format in their configuration files.
These classes provide a way for capturing the expressed concept as
pydantic models. This will enable adding the aforementioned modules to
`ngen.config`.