feat: CSList model- coerces a comma separated string of Vs into list[Vs]
Wrapper around a List[V] that coerces a comma separated string of Vs
(e.g. "1,2,3") into a List[V] (e.g. [1, 2, 3]). The `dict` method is
overridden to return the inner list of Vs instead of {"__root__": [Vs...]}.
As a result, using a CSList[V] as a field type in anther pydantic model
behaves as if it were a List[V] with the added functionality of coercing a
comma separated string into a list of Vs.

The features this model provides could easily be achieved using a field
level validator, so in some ways this is provided as a convenience. This
model is most useful when a field can be multiple types (union type). It
alleviates the need to implement a complex field level validator that
handles all possible union type members.