feat: model_vars mod. represent and resolve BMI module inputs & outputs
This module introduces functions and abstractions for representing and
resolving NextGen BMI module input and output variables. For example,
this feature set enables representing a set of BMI modules and
determining which outputs map to which inputs. Likewise, it enables
validating that a set of modules' output variables are valid under the
NextGen frameworks invariants.

TODOs:
- property resolve inputs that originate from the previous time step.