feat(ngen_cal): initial grew wolf optimizer implementation
This commit combines two files implementing the grey wolf optimizer as a pyswarm
sublcass.  The code from each as originaly authored by Xia Feng are included here
in the history.  These will be refactored and refined to fit into the general
ngen_cal architecture and assist in maintaining the implementation in the future.

Co-authored-by: hellkite500 <nels.frazier@noaa.gov>