refactor: add private `_optimizers` subpackage; move gwo there; export gwo in public `optimizers`
This refactor adds a layer of indirection to safeguard the `optimizers`
public api. The `_optimizers` subpackage provides a place for _both_
stable and unstable / experimental optimizers to live. Once an
optimizer's api and features are fleshed out, it can then be imported by
the `optimizers` subpackage in `__init__.py` effectively moving it into
the public api.