feat(ngen.cal): optionally run validation simulation w/ best params
NOTE: this needs to be refactored... and tested. This was pushed through
just to get thing working will not affect the code around it if this
feature is not used.

This feature enables running ngen with the best calibration parameters
after calibration has completed. A validation interval is configurable
under `model.val_params`. See ngen.cal.model.ValidationOptions for
specifics.